ovn: wait until node switch appears before creating the management port
Today, ovnkube-node checks the existence of the "ovn_host_subnet"
annotation to assure that ovnkube-master is up and running. This can
cause problems as the annotation was not cleaned up when one attempts to
deleting/recreating ovnkube-master deployment without recreating the k8s
cluster. Although this can be considered as a user error, it would
always be better to regard OVN DB as the source of truth instead and
check that.

Signed-off-by: Yun Zhou <yunz@nvidia.com>