Add option for host CNI client to configure the interface
Current ovn-k8s-cni-overlay CNI use client/server design to try
finish all the setup in cni server which running in ovnkube node
pod's container. In order to do that, container need run in previliged
mode or with SYS_ADMIN capablity in non-previliged mode. From security
perspective, SYS_ADMIN is mostly previliged mode, in order to avoid
that, this feature introduce a feature to let CNI binary which run
in host to configure the device.