Race between ovnkube-node POD and ovnkube-master POD
On addition of a node to a K8s cluster, the ovnkube-master process is
responsible for bringing up the node into OVN logical topology
(i.e., adding the logical switch for the node, assigning CIDR for it,
and so on) and finally annotates the node with the CIDR value it
assigned to it.

The ovnkube-node process on the node itself waits for all of this to
happen and then continues to ready the node  – management port creation,
gateway creation, and so on – for the scheduling of the pod.

Currently, the ovnkube-node waits only for the logical-switch creation
and continues with its tasks. However, it should also wait for the Node
annotation to show up before it proceeds.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>