make return arguments from GetNodeWellKnownAddresses similar
right now it returns router IP as IP/PLEN, whilst returns the management
port IP as IP. make the return arugments to be just IP so that the
callers can make use of the returned PLEN to construct IP with mask.

this change makes this function usable elsewhere in the code as well.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>