Move pod annotation under k8s.ovn.org namespace
The pod network info of IP, MAC, Gateway, and Routes are under
'ovn' annotation. We need to move it under 'k8s.ovn.org/pod-network'.

The changes assumes that the master is upgraded first. It continues to
write both the old/new annotation names to facilitate yet-to-be
upgraded ovnkube nodes. In the next release of ovn-kubernetes, we can
remove the code that adds `legacy` annotation.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>