GetK8sClusterRouter() function is expensive when it doesn't need to be
GetK8sClusterRouter() returns the OVN distributed router in the K8s
OVN topology. In a 400-node cluster, that function searches through
801 routers to find the distributed cluster router (the one with
external_ids:k8s-cluster-router set to yes). Initially, it was written
for use by clients such as ovnkube-node.

However, after the gateway creation move to master there is no need for
that function since the master already knows the name of the distributed
cluster router since it created it. So, remove that function and
directly use the constant that captures the name of the distributed
router.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>