add node systemID to the list of information passed from node to master
OVN Gateway logical resources are now being created on the master. To
create a L3 gateway, we need the node's system ID. We are retrieving
this information in the master by using

   `ovn-sbctl find chassis hostname=<nodename>`.

In a large cluster with 100s of kubernetes nodes and busy OVN SB DB,
sometimes we timeout while getting this information.

It is better for each node to pass the system ID through node annotation
along side with other information it currently passes for gateway
creation.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>