Fix parsing of gateway-options
daemonset.sh uses awk $2 with "=" delimiter so it misses the rest of the options if there is a = in it.
For example: daemonset.sh with --gateway-options="--gateway-interface=eth2 --gateway-nexthop=172.17.20.1"
ends up with: ovn_gateway_opts: --gateway-interface
while it should be:
ovn_gateway_opts: --gateway-interface=eth2 --gateway-nexthop=172.17.20.1
Use cut to take all the variable text.

Signed-off-by: Shahar Klein <sklein@nvidia.com>