scale: ascertain gateway readiness by checking OpenFlow rules
With 400+ odd nodes, the current GatewayReady() function is not
scaling. The ovn-nbctl calls are timing out. When we have a way to find
out that the data path for the L3Gateway is ready by checking
for OpenFlow rules on the integration bridge we should make use of it.

Adding SNAT rules is the last thing we do while building the logical
topology. So, check for the SNAT rule in table 65 in the integration
bridge

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>