do not delete integration bridge as part of ovnkube-node cleanup
when we delete ovnkube-node, to say fix some issue or it met a panic,
we delete the integration bridge. this inturn deletes all the POD's
ports on the integration bridge. When the ovnkube-node comes back up,
the integration bridge is created but not the ports on it. as a result
all the PODs are disconnected from network.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>