Clustered database (aka RAFT) support for OVN DBs
This commit implements the clustered database as a statefulsets with
replicas set to 3. The PodManagementPolicy of OrderReady is too
restrictive for how RAFT works, so instead Parallel policy is used.
The existing ovnkube-db service acts as the headless Service for the
statefulset. The pods are scheduled on the nodes with labels
`ovn.org/ovnkube-db=true`.

Implementation notes:

1. ovnkube-db-0, the first pod in the statefulset creates the initial
cluster DB. If an existing standalone DB is present, then it converts it
into the clustered DB.

2. ovnkube-db-1 ... ovnkube-db-{n-1} pods wait until the ovnkube-db-0
has successfully started and then only join the cluster.

3. ovnkube-db-{n-1}, the last pod in the statefulset will ensure that
all of the pods in the statefulet are up and running and then only
publish the ovnkube-db endpoints to the ovnkube-db service.

4. Like before, all of the ovnkube-nodes and ovnkube-master are waiting
on the endpoints to be published. Once the endpoints are published, they
proceed further.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>