refactor Link{Addr|Route}Add API and add Link{Route|Neigh}Exists API
Link{Addr|Route}Add API was deleting the existing entries and adding
the new entries. During the review it was suggested to split this into
two functions. This commit takes care of it.

It also adds two new APIs Link{Route|Neigh}Exists that will be used
by health check routines in future.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>