add a metric to capture the number of times a resource is being udpated
this metric will give insight into how often a UpdateFun for resource,
(pod, services, endpoints, and so on) that we care about gets called.
very recently we found an istance of an operator going rogue and calling
node updates every 5s, and the ovnkube-master CPU peaked since it was
constantly executing Node UpdateFunc.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>